cancel button works
only slightly off-putting thing is when we start the app,
an base-class Item object populates the view pane.
If user starts editing a new item and clicks 'cancel',
the view switches back to the view pane but the populating Item got
replaced by an empty Item of the class that was just edited.

This happens because upon switching to the edit tab for new item
editing, the view_var gets filled with a blank Item of the appropriate
class.